End result (#1)
* Refactor, added some queries, now possible to track modification date

* bugfix

* Notification support added

* Date format change

* Now ignoring requests where we try to create markers for the same location

* bugfix

* Refactor and grading

* Rename refactor

* User support added, further renaming, scheduled operations currently off

* Switched to proper DB

* Dev switch introduced

* User reliability logic + refactor (Lombokisation)

* Logic fixes

* Logic fixes

* Endpoint harmonisation

* Unused test removal

* Update application.properties

+++ src/main/java/pl/herfor/server/data/LoadDatabase.java
@@ -5,27 +5,37 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import pl.herfor.server.data.objects.AccidentType;
-import pl.herfor.server.data.objects.Marker;
-import pl.herfor.server.data.objects.MarkerProperties;
-import pl.herfor.server.data.objects.SeverityType;
-import pl.herfor.server.data.repositories.MarkerRepository;
+import pl.herfor.server.data.objects.Report;
+import pl.herfor.server.data.objects.ReportProperties;
+import pl.herfor.server.data.objects.User;
+import pl.herfor.server.data.objects.enums.Accident;
+import pl.herfor.server.data.objects.enums.Severity;
+import pl.herfor.server.data.repositories.GradeRepository;
+import pl.herfor.server.data.repositories.ReportRepository;
+import pl.herfor.server.data.repositories.UserRepository;
+
+import static pl.herfor.server.data.Constants.DEV_SWITCH;
 
 @Configuration
 @Slf4j
-public
-class LoadDatabase {
+public class LoadDatabase {
 
     @Bean
     @Autowired
-    public CommandLineRunner initDatabase(MarkerRepository repository) {
-        // TODO: remove to retain persistency
+    public CommandLineRunner initDatabase(ReportRepository repository, GradeRepository grades, UserRepository users) {
+        if (!DEV_SWITCH) {
+            return args -> log.info("Database init ignored (no dev switch).");
+        }
+        grades.deleteAll();
         repository.deleteAll();
-        MarkerProperties properties = new MarkerProperties(AccidentType.METRO, SeverityType.YELLOW);
+        ReportProperties properties = new ReportProperties(Accident.METRO, Severity.YELLOW);
+        User user = new User();
+        user = users.save(user);
+        User finalUser = user;
         return args -> {
-            log.info("Preloading " + repository.save(new Marker(52.3501, 20.8600, properties)));
-            log.info("Preloading " + repository.save(new Marker(52.3801, 20.8900, properties)));
-            log.info("Preloading " + repository.save(new Marker(52.4001, 20.89100, properties)));
+            log.info("Preloading " + repository.save(new Report(52.3501, 20.8600, properties.withSeverity(Severity.RED), finalUser)));
+            log.info("Preloading " + repository.save(new Report(52.3801, 20.8900, properties.withAccident(Accident.BIKE), finalUser)));
+            log.info("Preloading " + repository.save(new Report(52.4001, 20.89100, properties.withSeverity(Severity.YELLOW), finalUser)));
         };
     }
 }
